[AM] Rename StopApplication into DestroyEXC
This belongs to a series of patches intended to cleanup some code
by aligning RTLIB concepts with the Barbeque internals.

This patch provides the renaming of StopApplication with DestroyEXC.
Indeed, internally to Barbeque, each single and "atomic" computational unit
which could be associated to resources is named "Execution Context" (EXC).
Each EXC could be associated to a specific recipe and each Application could
be composed or use multiple different EXC.

Once an EXC is no more needed, we could release it an all the resources
associated to it by using this new method.

NOTE: we moved the clean-up code for the Application class into its destructor.
Indeed, for the moment, once an EXC is destroyed we do not move it into
the FINISH state (which is thus an "ephemera" state) but instead we
prefer to cleanup it and destroy. If the FINISH state should be needed,
e.g. to keep track of some information or statistics on EXC usage, than
we could change this design decision.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>