[AM] Rename DisableApplication into DisableEXC
This belongs to a series of patches intended to cleanup some code
by aligning RTLIB concepts with the Barbeque internals.

This patch provides the renaming of DisableApplication with DisableEXC.
Indeed, internally to Barbeque, each single and "atomic" computational unit
which could be associated to resources is named "Execution Context" (EXC).
Each EXC could be associated to a specific recipe and each Application could
be composed or use multiple different EXC.

An application could register and use multiple different EXC, but only
enabled ones are considered for scheduling by the Barbeque RTRM.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>