[Application] Update EXC Disable
This patch update the support for disabling an EXC.
Depending on the current EXC state, different actions should be take:
- READY EXC are simply marked disabled
- RUNNING EXC require a synchronization to release all the resources.

NOTE: this implementation assumes that, once an application exit the system,
its resources are auto-magically released by the ResourceAccounter.
This is based on the hypothesis this component exposes a "transaction based
update of the current system state.
Refers to discussion related to tickes #26.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>