[Application] Scheduling state code refactoring
This patch remove the usage of "curr_sched" and "next_sched" by replacing them
with a single "schedule" structure which is used to hold information from
the current application state.
This data structure provide also information to support synchronizaiton:
- next_awm keep track of the optimizer selected AWM
- preSyncSTate keep track of the application state before entering the
SYNC one
Both these info are useful for the synchronization module to know which
configuration should be enabled for a re-configuring applications and which
is the correct final state.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>