[Application] Added verbose state
This patch add a verbose representation of both current states and
synchronization states which could be used for better logging.

NOTE: A final "_COUNT" value has been added, which is not a valid state
but it is used to size the string vectors.
Moreover, this values could be conveniently used to represent a "NOT valid
state" (i.e. an error condition), for example by methods which are
required to return a state.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>