[Application] Scheduling/Sync methods completed
ScheduleRequest():
The method check if the resource set passed as argument 2 is allocable. If
so the set is bound into the working mode specified as argument 1. Then
Reschedule is called passing the working mode (and eventually the resource
view token). If Reschedule() returns with success, set the working mode
as the "next" of the Application.

SyncRequired():
This is called for "RUNNING" Applications/ExC with the aim of determine the
"synchronization state" between RECONF, MIGRATE, MIGREC, SYNC_NONE. Last
one is returned in case of Application/ExC rescheduled in the same AWM
using the same set of resources.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>