[Application] Bugfix RequestSync
This patch bugfixes a SEGFAULT generated by the RequestSync, which was passing
a new AppPtr_t to the AM, thus producing a release of the Applications as soon
as the AM::SyncRequest returned. :-/

Now we exploit the UID support to get a valid AppPtr_t from the AM and pass
it for the SyncRequest.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>