[Application] Removed TERMINATE sync state
The TERMINATE sync state is a no sense since sync states are assigned only
by the optimizer which however never kills an EXC but eventually, if resources
are not available, could request to BLOCK it by switching the EXC on READY
state.

Termination of an EXC could be requested only by the contolling application
or by other Barbeque modules. In any case, EXC resources should be released
as soon as possible and a notification delivered to the ResourceManager to
decide whether re-run an optimization.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>