[SchedPolicy] Added DISABLED/FINISHED App check
During the scheduling computation, a check is made in order to determine
if the Application/ExC is still enabled. Indeed it may happens that
right after the start of the SchedulingPolicy an Application has
launched a Disable or Stop command. Accordingly the Application should
be skipped. The check is done before each AWM metrics computation and
repeated in the selection phase.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>