[AM] Added Sync Status MUTEX
This patch add a mutex protection for each Synchronization queue.
A new set of methods are also provided to properly access sinchronization
queues, by looking them, and add or remove elements.

NOTE: according to this patch, EXCs are never moved from a sync queue to
another, but they can only be added or remove.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>