[Application] Bugfixed SyncRequest state notification
The AM::NotifyNewState() should be before the actual A::SetState() otherwise
the notifications assert fails. Indded, within notify a check is performed
to verify that the current state and the next one are different. This is
required to know how to move EXC between queues.

NOTE: this patch requires a testing and review, especially regarding "corner"
cases such as EXC destruction.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>