[ApplicationManager] Bugfix SyncRequest
Once we enter a sync request, the application status has already been updated
to be in SYNC state. Indeed, according to the new design, the AM::SyncRequest
is used just to notify the AM that everything has been properly setup for
a synchronization to happen.

This method is called right now just by  A::RequestSync, which actually calls
it right after the EXC status has been updated and notified to the AM.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>