Added RPC FIFO plugins module
This patch add an initial implementation for a simple FIFO based module
implementing the RPCChannelIF. This is intentionally a simple
(and still incomplete) implementation which is based on blocking calls
and memory less message reception.
A proper proxy class is expected to provide support for received message
queuing and processing.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>