[ResourceScheduler] Added first implementation
This patch add the first implementation of the ResourceScheduler class.
This class is used by the ResourceManager module once a scheduling of
resources is required. Resource scheduling is triggered by events, which
could be:
1. EXC requesting a new working mode
2. EXC termination
3. resources availability changes
These events are served in decreasing priority by the ResourceManager,
i.e. resource availability changes have higer priorities over EXC start.

This patch (partially) integrate just the lowest priority event: an
EXC requesting a new working mode. This event trigger a call to the
ResourceScheduler which should decide if a new schedule could be attempted.

However, two main pieces are still missing on the current implementation:
1. the usage of a scheduling activation policy (ticket #19)
2. a proper statistics collection support (ticket #20)

Signed-off-by: Patrick Bellasi <derkling@gmail.com>