[TestPlatformData] Add support for test platform data
This patch add a new core module which could be used to load a test platform.
The module provides a set of command-line (or configuration file) options
which could be used to define:
1. the number of clusters;
2. the amount of cluster-shared memory, in [MB];
3. the amount of PEs per each cluster.

All these information are used to properly configure the ResourceAccounter
with a valid TEST platform. By default, the loaded platform is a mimics of
the Numasauro setup, i.e. 3 cluster, with 4 PEs each and 8GB of cluster
share memory .

This support is enabled only if the BBQUE_TEST_PLATFORM_DATA cmake option
is enabled (which is by default right now) and could be used also by test
plugins, since the configuration is loaded before checking for them.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>