[ResourceAccounter] Synchronized mode support
The support to the synchronized resource acquisition have been added.
This feature should be exploited by the SyncronizationManager in order
to update the system resources state (accounting).

To do that the first step is calling SyncStart(), which open the
syncronized session. This method init a new resource state view by
starting from the resources used by the Application/ExC in RUNNING
scheduling state.
Next step, by the SynchronizationManager side, is to call a
SyncAcquireResources() for each Application/ExC that needs to be
reconfigured or migrated (or both). This call determines the resources
acquisition by the Application/ExC.
Once done the SM must commit the session, by calling SyncCommit(), which
makes changes effective and close the session. If something goes wrong,
a call to SynAbort() cancel the session, without bringing changes to the
system resources state.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>