[RandomSchedPol] Added Random scheduler policy
This patch add a simple and dummy random AWM scheduling policy.
Such a policy is useful just for debugging and stress test of other
Barbeque framework modules.

The policy simply select a random AWM for each application in READY
or RUNNING state. The selection in based on a uniform RNG working on
range [0..100]. Thus such a policy should be just fine to stresstest
recipes with up to 100 working modes.

The random scheduling policy is the default plugin loaded when
Barbeque is compiled in DEBUG mode and no other policy is specified by
the configuration file. Instead, compilation of this policy is
disabled when building in RELEASE mode.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>