[SyncPolIF] Added plugin interface for Sync Policy
This patch provides the interface definition for the development of
synchronization policy plugins.

This interface should be supported by synchronization policy plugins.
Such a policy is used by the SynchronizationManager (SM) to identify the
set of EXCs which should be synchronized. A policy is required just to select
the EXCs and than pass back the selection to the SM to perform the actual
synchronization.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>