[SynchronizationManager] Added first implementation
This patch provide a first implementation of the SynchronizationManager
module. This is a core component of the Barbeque RTRM framework which
is in charge to identify synchronization needs and actually perform them.

A synchronization could be required each time a new scheduling has been
identified by the SchedulerManager (now named ResourceScheduler) component.
This component is in charge to communicate the new schedule to all the involved
EXCs, by properly managing and assisting the reconfiguration of each of them.

The decision to actuate a scheduling, thus starting a new synchronization,
is completely on charge to the SM component, which is assisted on this role
by two policies:
1. the Synchronization Policy, which is in charge to select the order on which
EXCs must be re-confgiured
2. the Synchronization Decision Policy, which is in charge to decide when a
synchronization could be actually performed

NOTE: this is just an initial implementaiton of this component, providing the
basic support to link it with the ResourceManager and a SynchronizationPolicy.
Specifically, this implementation is still missing all the EXCs communicaiton
code.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>