[SyncPol_SASB] Add first implementation
This is the first implementation for the "Starvation Avoidance State Based"
(SASB) synchronization heuristics. The goal of this synchronization policy
is to re-configure EXCs trying to avoid contention in resources acquisition
thus synchronizing before the contexts which are more likely to release
resources (e.g. blocked ones) and than these which are more likely to acquire
resources (e.g. starting ones).

Of course, this is not a "dead-lock safe" policy regarding resources access,
however it should provide a relatively good starting point for experimenting
with the Barbeque RTRM.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>