[SchedPol_Random] Bugfixed Apps iterator segfault
The iterator on EXCs faults if the application is removed from the iterated
queue. This is certainly due to the copy of the end() iterator at the beginning
of each iteration... however, the segfault seem to persist even if we fixes
this initialization. Thus, right now we fix the problem the hard way by making
a copy of the apps vector before starting the scheduling iteration.

Probably this solution does not completely solve the problem and it is also
inefficient. A better solution is under investigation.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>