[Application] Bugfixed assert checks on NotifyNewState calls
This patch move the call to ApplicaitonManager::NotifyNewState() within the
SetState() where it is more simple and efficient to consider all the possibile
states.

Entering a SYNC state, the SyncState must be configured before notifying the
new state, since this information is required to move the EXCs into the proper
queue.

Entering a stable state, the state must be updated after the notification,
since the current state must be knowen to remove the EXCs from the correct
queue.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>