[Application] Bugfixed call to AM::SyncRequest
If an AM::SyncRequest() fails, the AWM could still be accepted. Indeed, the
resources are available and the EXC has already been moved into the
proper sync queue.
What is failing there is the notification of a possible Synchronization,
which according to the current design it is a quite remote event.
Thus, this should not be a main issue to justify the reset of the state
required for the EXC.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>