[WorkingMode] Null pointer fixed in BindResource()
A check upon the UsagesMapPtr_t passed was missing. Thus a shared
pointer to null brought to a segmentation fault.

The problem has been overcame by allocating a new UsagesMap_t if the
pointer is null.  Moreover this approach disambiguates the usage of the
method. Indeed from the caller perspective it wasn't clear if the
pointer passed should point to a map allocated yet or if the map is
going to be allocated in BindResource(). Now, both usages are supported.

This patch also makes a "transparent" refactoring of the method
signature. The char string argument (set to NULL as default value) has
been removed. It was used to store the path of a resource eventually
skipped in the binding process. However this is a fatal error condition,
and thus the code setting that string has been replaced by an assert.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>