[RTLIB] Switch from GWM to SCHEDULE
This patch remove the misleading RPC_EXC_GWM call from the RTLIB and
replace it with the RPC_EXC_SCHEDULE request. Indeed, applications use the
RTLIB::GetWorkingMode() to get the ID of the current AWM to be used.
If such a valid AWM is not know by the RTLIB, a SCHEDULE request is issued
to Barbeque while the apps will remain supsended in its call waiting for
either a response from the RTRM or a timeout.

TODO: add a timeout version of the RTLIB::GetWorkingMode() which allows
an application to specifiy the maximum amount of time it could wait for
a valid AWM being assigned.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>