[SM] Added ResourceAccounter SyncMode session
Once a scheduling has been defined, SynchronizationManager should
perform the resource acquisition for each application/ExC to deploy on
the platform. This is done by opening a "synchronized mode" session with
ResourceAccounter, and requiring the resources. If everything goes fine
SM can "commit" and thus close the session. Otherwise it can abort it.

This patch add this interaction between SynchronizationManager and
ResourceAccounter.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>