[RPC] Using uint8_t to represent RTLIB_ExitCode
Using an uint8_t instead of the default int datatype is more efficient in
terms of RPC messages space required. However, we must be aware to not
exceed the representation capabilities with the definition of the
RTLIB_ExitCode enum, thus an assert should be added on each statement
formatting an RPC message.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>