[SM] Add support for Sync-, Do- and Post-Change
Integrate the initial implementation for the SyncChange, DoChange and
PostChange synchronization protocol messages provided by the AM.
This implementation should allows to properly deliver messages to BBQ
registered applications and collect back responses from them.

The synchronization protocol is based on four main calls with different flawors
according to their purpose:
1. PreChange  - Sync or Async, Response required (expected sync latency)
2. SyncChange - Sync or Async, Response required (noting defined right now)
3. DoChange   - Sync, Response NOT required (this is just a notification)
4. PostChange - Sync, Response required (noting defined right now)

Please note that, based on the current implementation, a call to the
AM::SyncCommit() change the apps queue the SM::PostChange() is looping on.
To avoid problems with iterators we prefer to relay on queue::begin() to
get a reference to the new queue head at each now iteration.
However, this approach risk to produce and endless loop if the previously
synchronized application is correctly remove from the sync queue.
This condition should be better virefied and properly handled.

Moreover, the current implementation is missing any kind of interaction with
a (still to be defined) synchronizaiton policy, which should evaluate the
required reaction on anomalous conditions, e.g. an apps not reaching the
expected sync point when a SyncChange is issued.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>