[RTLIB] Add SyncP logic
This is the first working implementation of the Synchronization Protocol logic.
The four main methods have been implemented to properly manage the EXC state,
locking and notification. This required also a properly integration with
the RTLIB::GetWorkingMode call.

This is the expected main workflow:
1. BBQ => PreChange:
   enter "EXC SYNC" mode
   invalid current AWM
   set reconfiguration action required
   set next required AWM
2. APP => GetWorkingMode
   notice the sync mode
   enter "AWM WAITING" mode
   sleep on waiting for an assigned AWM
3. BBQ => SyncChange
   report on being into "AWM WAITING" mode
4. BBQ => DoChange
   set valid AWM
   notify the APP
5. GetWorkingMode => APP
   get the new required AWM
   switch configuration according to the required action
6. BBQ => PostChange
   wait for reconfiguration to complete
7. APP => GetWorkingMode
   (after reconfiguration completed) check for the assigned AWM
   get quthorization to run on that AWM
   enter "SYNC DONE" state
   notify the RTLIB of reconfiguration compelted
8. PostChange => BBQ
   report on reconfiguration completed


TODO: this basic workflow still need to be updated to properly manage exceptions
such as a reconfiguration tacking too much time. Indeed, in this case for
example a timed-out wait should be used to report the RTRM about this condition.


Signed-off-by: Patrick Bellasi <derkling@gmail.com>