[AP] Pass-by-Copy the SyncP responses
This patch replace usages of pass-by-reference with usages of pass-by-copy
for the shared pointers to the synchronization protocol messages.
This is required to have a working handling of shared pointers especially
considering the spawning of new threads, where passing the argument by
reference seem to be not stable.

TODO: try to revert this patch on top of a stable tree to verify if it is
really needed.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>