[ResourceAccounter] Fixed bug in GetView()
The hash function returning the token for the resource state views
didn't worked as expected. Whenever a test with more than one
synchronization was launched an assert in RA was raised.
The problem came from the fact that the sync session in RA continued to
get the same rv token for each session, thus resource booking was
performed on a "dirty" resource state view.
Since the template parameter has been changed from "const char *" to
"std::string" in std::hash<>() the method GetView() seems to work
correctly.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>