[RM] Decouple pending event set from notification
Marking an event as pending could be done even while the RM control loop
is running. Indeed, doing so we have a chance to handle multiple events
during the same run.
However, to ensure that an event is actually noticed, we need to sync with
the start of each new control loop. Thus, after the events has been marked
we exploit the pendingEvts_mtx to wait for a control loop run to finish
and than we notify the control loop for a new run to start.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>