[Application] Added current AWM check in Terminate()
Once an Application/EXC stops its execution, a check on the current AWM
is required before ask for resources release. Indeed it may happens
that such stop occurs before a synchronization phase is ended. In such
case, if the Application/EXC is READY, it hasn't any AWM assigned and
thus doesn't hold any resource. Calling a release on ResourceAccounter
would be useless because RA won't be able to find any reference to the
Application.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>