[RTLIB_TestApp] Updated based class termination
The termination of the application has been re-written to be more
Barbeque safe. Indeed, we ensure the EXC is un-registered before trying to
stop the control thread.
This is safer, since right now, on some extreme corne cases, we are still not
able to cleany stop the control loop. In these cases, we are however shure
the EXC has released its resources.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>