[RTLIB] Add support for SyncP TIMEOUTs
This patch add TIMEOUT support to all the SyncP calls to avoid to remain
blocked on a not responding application.

One a SyncP call send a command, it wait for a response from the application
up to a (right now hard coded) timeout of 200ms. If a response is not received
by this time, the application is considered misbehaving and the EXC is
DISABLED.

The proper reaction to a TIMEOUT should be policy defined. Thus, this code
is expected to be moved ASAP into a proper policy plugin. This pluging will
be in charge also to define, perhaps dinamically, the maximum allowed timeout
value for each SyncP call.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>