[APP] Add MUTEX protection for query methods
This patch introduces mutex based protection on methods provided by the
Application class to query internal state. This is required to ensure reading
integrity since the internal status depends also on modifications of
ApplicationManager queues.

Moreover, since the tight cooperation among the Application and the
ApplicationManager, the previously used mutex for the schedule state locking
has been replaced by a recursive_mutex, which allows keep the status mutex
recursively by the same thread. Indeed, actions on the Application objects
are always triggered by the ApplicationManager on behalf of a command
session thread originated to satisfy an application or Barbeque request.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>