[AP] Bugfixed race condition on SyncP responses reception
It could happens that a response is received even before the *ChannelRecv
methods had the chance to be scheduled. This has been noticed for instance
on single core platforms, where sending a message to the application
triggers a rescheduling which allows the application itself to process
the message and generate the corresponding response to Barbeque before
the *ChannelRecv code is rescheduled to process it. In this case *ChennelRecv
methods falls into an endless wait.

This patch solves this race condition by exploiting the message buffer pointer
as a guard variable to know if the response has been already received, otherwise
(an only in this case) the *ChannelRecv code sleeps waiting for the response.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>