[AP] Bugfixed race condition on dispatcher thread sturtup
It could happens that, if the thread creating the ApplicationProxy calls also
the Start() method befor even the Dispatcher thread has had a change to be
scheduled. In this case, the Dispatcher will sit into an endless wait for
the Start(). This patch solve this race condition.

A boolean flag has been added to define if the status of the Dispatcher thread.
This flag is set to TRUE once the thread is started by a call to Start().
The flag status is than tested by the Dispatcher thread in order to know if
it should sit down waiting for an actual start, or it could go one.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>