[RPC] Improved Disabling/Unregistration management
This is a set of modifications which exploits the REGISTER and ENABLE flags
to manage the asynchronous disabling of an EXC while the ControlLoop is
running. By properly managing these flasg we are able to clean exit from
an EXC event when it is waiting on a Barbeque event (e.g. waiting on a
GetWorkingMode).

Signed-off-by: Patrick Bellasi <derkling@gmail.com>