[PluginManager] Forced debug disable
Debug sentences for this module are almost useless but also very boring at
run-time, especially on highly dynamic scenarios where EXC enters and exit
the system with higher frequency.
Thus, by default we force DEBUG off for this module.

TODO: A better debugging configuration file should be provided, in the style
of the one used by BeRTOS to selectively eforce-off debugging for single
modules. Indeed, this will allows at leas to have a single point of check for
a proper configuration of the debugging support.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>