[RM] Add Optimization trigger on EXC disable
When an EXC is disabled, a new Optimization cycle is tirggered.
A new provided Optimize method ensure to call both scheduling and synchronization
just when there is some opportunities, i.e.:
- Scheduling: at least one applications READY or RUNNING
- Synchronization: at least one application on SYNC

TODO: once an EXC is disabled we wait 500ms before triggering the optimization.
This is a dummy heuristics to avoid optimizations runs when multiple EXCs are
being released almost at the same time.
Of course, a better policy should be defined!

Signed-off-by: Patrick Bellasi <derkling@gmail.com>