[Application] Fixed SYNC/BLOCKED case in Unschedule()
If an application/EXC schedule requested fails, due to ResourceAccounter
response, a state SYNC/BLOCKED is set. If the scheduling policy recall a
ScheduleRequest() for this application/EXC, method Unschedule() is called.
Here an abort is raised due to the fail of the assert (state == RUNNING).

This behavior has been fixed by checking if the Application is in
SYNC/BLOCKED state. If true, Unschedule() returns without reaching the
assert.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>