[YamcaSchedPol] Added clusters full check
To increase scheduling efficiency, if a cluster doesn't contain available
processing elements it is skipped.
This control is triggered whenever we are checking for the availability of the
PEs required by an AWM. If ResourceAccounter returns 0 then a proper exit code
is returned (SCHED_CLUSTER_FULL). The fullness state is thus tracked by a vector
of boolean, having size equal to the number of clusters.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>