[Application] ScheduleRequest() on SYNC/BLOCKED
If the Application is in state SYNC/BLOCKED, a ScheduleRequest() can be still
performed. Indeed it commonly happens that if a scheduling/allocation fails, the
scheduling policy want to try again by specifying a new resource set.

This was not possible before and now it has been fixed by resetting the state
from SYNC/BLOCKED to the previous one (it may be READY/SYNC_NONE or
RUNNING/SYNC_NONE reasonably), before doing the request.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>