[AM/App/RA] Added commit for RUNNING applications
If an Application/EXC has been scheduled for continuing to run in the previous
AWM, using the same resources, we need to commit this state even if this is not
a SYNC one. This is required by the SchedPolicy logics.

This is performed by ResourceAccounter in the "sync session", wherein running
applications initialize the next system state view. Here RunningCommit() of
ApplicationManager is called for each of this applications.

By its own side RunningCommit() forwards the call towards ScheduleContinue() of
Application class. What this method does is a simple reset of the next AWM, once
checked that the Application is in RUUNING state and that current and next AWM
are the same.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>