[BuildSystem] Add Doxygen documentation
This path add support for the compilation and installation of the
Doxygen documentation. The documentation is NOT automatically build by
the all target, an explicit "make doc" is required. However, if build,
the install target deploy the documentation into the folder specified by
the BBQUE_PATH_DOCS cmake configuration variable.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>