[YamcaSchedPol] Use CheckSkipConditions() to skip
Method CheckSkipConditions() has been added to refactor some replied
lines of code, and thus improve efficiency and readability.
The method embeds the skipping conditions checks on a given
Application/EXC. It is call both during ordering and selection step, and
returns true if one of the conditions (explained in commit 9a0e4ea47eb3)
are verified.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>