[RS] Rename ResourceScheduler into SchedulerManager
For consistency with the overall framework design, this patch rename the
ResourceScheduler into SchedulerManager (SM). Indeed, like the
SynchronizationManager the SM is a proxy to trigger a (resource) scheduling
policy.

For consistency, the namespace of this Barbeque component has been updated
to be "bq.sm"; thus, in order to avoid collisions, the namespace of the
synchronization manager has been updated to be "bq.ym" and from hencefort
the SynchronizationManager will be referred as YM module.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>