[Utils] Split resource related stuff from generic utility
This patch move the resource related stuff, previously defined within the
utility header, into a self contained header which could be included by its
client modules. This is required in order to allow the deployment of the
utility header, as well as the timer header, with the RTLib so that they could
be exploited by applications as well.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>