[RTLIB_RCP] Add UnregisterAll method
This patch provides a new public method which allows to unregister all the
so far registered EXCs. This method could be used by the RTLib distructor to
ensure that all the resources assigned to this application are properly
released once the application is exiting, thus making the library more
robust against lazy application coders.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>