[RA] SyncAcquireResources() refactoring
Signature of SyncAcquireResources() has been refactored by removing usages
map argument. Thus the application pointer is the only argument to pass.
The method now retrieve the resources usages map from the Application next
AWM. If this is missing, a fatal error is returned.

This patch aims to simplify the method usage and improve readability.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>