[AM/Application] Added sync abort support
ApplicationManager: Added SyncAbort() method for aborting scheduling of an
Application/ExC. This should be called (by SyncManager) whenever
ResourceAccounter signals a fail during the synchronized resource
acquisition for an Application/ExC.
Note: The Application/Exc MUST be in a SYNC state.

Application: Added ScheduleAbort() to notify the Application/ExC about the
abort of its scheduling and thus synchronization. The method reset the AWMs
settings and set the state to READY.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>