[SM] Migrate the ILES iterators
This introduces the usage of in-loop erase safe iterators on the
SynchronizationManager class.
This required a main modification on the current implementation since, in order
to grant safety access on mutable queues, right now we cannot realy anymore
on policies returning a queue pointer but just the queue ID to process.
Thus, the new plugin interface has been updated to require the ploicy to
return just the SYNC queue to synchronized. Than, the SynchronizationManager
core exploits an ILES iterator to visit the queue and properly sync all the
applications returned.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>