Add In-Loop-Erase-Safe (ILES) Iterators support
This patch adds support for iterators which are safe to erases while visiting
the container.

The support is based on the newly added AppsUidMapIt  class, which is an
iterator wrapper object which is used to implement safe iterations on
mutable maps, where an erase could occours on a thread while another thread
is visiting the elements of the same container container.

A proper usage of such an ILES iterator requires to visit the container
elements using a pair of provided functions "GetFirst" and "GetNext".
A pair of such functions is provided for each queue exposed by the
ApplicationManager.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>