[RTLIB_RPC] Keep track of Control Thread PID
This patch keeps the PID of the control thread so that it could be exploited
by the Barbeque to enforce decisions on each EXC (e.g. to kill it if required).

Moreover, the initialization of registered EXC is now done using a struct
constructor, which endure a proper initialization of all the required
EXC parameters. Especially, both the map of registered EXC and the GetNextExcID()
are now static (i.e. class defined).

Signed-off-by: Patrick Bellasi <derkling@gmail.com>