[AEM_CPP] Bugfixed default onRun termination
The default implementation of the onRun should terminate after a pre-defined
number of cycles. However, due to a misused static countdown, only one
control thread was terminating in this buggy implementation.

To solve the bug, this patch provides also a more generic approach.
A cycles counter as been added, which is used both to termiante the EXC on
the default implementation but also to count the number of onRun executions.
This has been conveniently used on logging to improve readability.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>