[RTLIB_RPC] Bugfixed (Ticket #52) BLOCKED synchronization
This patch solves a bug with synchronization of BLOCKED applications.
Once an application should be blocked, e.g. due to unavailable resources,
a synchronization is triggered but the EXC has not an AWM assigned.

In this case, the RTLib will ignore the AWM value defined by the RPC message,
since the BLOCKED event is also defined. Once this happens, the EXC is marked
as blocked (thanks to a new added status flags) and the WaitForWorkingMode
returns thus thus letting the EXC Control Loop of the application to move
into the a GetWorkingMode, waiting for more resource being available.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>