[RTLIB_RPC] Add first statistics collection and reporting support
This patch add a new AwmStats struct which is embedded by each gregistered
EXC to collect statistics regarding the usage of each AWM.
Statistics for an AWM are collected only if an AWM is used at least one time.
For the time being we collect, at the application side:
1. number of times each AWM as been used
2. number of processing cycles executed on each used AWM
3. processing time [ms] spent on each used AWM
4. last but not least: an estimation of the synchronization perido for
each AWM

The synchronization period is computed so far simply using a simple Cumulative
Average (CA), using a predefined number of samples, which is updated at each
sync point (i.e. GetWorkingMode call time).

A set of methods allows to properly update the metrics of interest, by using
the EXC specific Timer, and have been integrated with the RTLib API to
match the AEM control loop defined by the BbqueEXC.

Moreover, a report on such metrics is logged at each EXC disabling time.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>