[RTLIB_RPC] Add sync delay compuatation
This is the first implementation of the Synchronization Delay compuatation.
The GetSyncLatency() now exploits the Boost::Accumulator statistics to
provide an estimation of the next synchronization point. Such estimation
is straightforward, we assume NORMAL distribution for sync time samples
and thus we estimate the sync time as:
      sync_time = average(samples) + 2 * std_deviation(samples)
Then the expected sync delay is computed by accounting for the time since
last sycle started and rounding up to the next millisecond.
This is the expected latency returned to Barbeque.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>