[SyncPol_SASB] Add first implementation for sync time estimation
This is the first implementation of a (dummy) synchronization point
estimation, which is based on the collection of the maximum reported
synchronization point by all the applications involved into a synchronization
cycle.

We exploit the GetApplicationQueue to know when a new synchronization cycle
is starting, thus resetting the maximum latency.
Each time an application has notified its latency, within the CheckLatency()
we simply update the current value for the max latency. This value is
reported once required by a call the EstimatedSyncTime method.

Of course:
1. the max_latency should be time-corrected considering the time instant
on which has been reported;
2. a validation policy should be implemented: not all latencies could be
acceptable, such a policy should consider also the specific application
asserting the latency.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>