[SM] Avoid resource acquisition on disabled applications
This is a best-effort patch which avoid resources acquisition on EXCs which
have been disabled during the SyncP_PostChange call.
Please note that this patch is not completely safe since the EXCs could be
disabled right after this call. The SyncAcquireResource method should take
care of this case and use a proper locking on papp in order to avoid the
EXCs being disabled while resources are acquired.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>