[AM] Bugfixed abort on DISABLED applications
It could happens, due to the fine granularity of EXCs disabling support, that
an EXC is disabled during a SynchP run. In this case the EXC is correctly
removed from sync queues. A proper check on this condition is already handled
within the ScheduleCommit call while a lock on schedule_mtx is hold.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>